backend/requirements.txt
alembic>=1.12.1
bcrypt>=4.0.1
beautifulsoup4>=4.11.1
fastapi>=0.104.0
fastapi-pagination>=0.12.4
passlib>=1.7.4
psycopg2-binary>=2.9.3
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose>=3.3.0
python-multipart>=0.0.6
SQLAlchemy>=2.0.0
SQLAlchemy-Utils>=0.41.1
starlette>=0.27.0
url-normalize>=1.4.3
uvicorn>=0.22.0
gunicorn>=21.0.1
aiofiles>=23.2.1
asgiref>=3.7.2
cryptography>=41.0.3
python-dotenv>=1.0.0
redis>=4.5.4
celery>=5.2.7
pytz>=2023.3
pydantic[email]
fastapi_pagination
starlette-admin